LayoutTests/storage/indexeddb/mozilla/readonly-transactions.html

<!DOCTYPE html>
<!--
  original test: http://mxr.mozilla.org/mozilla2.0/source/dom/indexedDB/test/test_readonly_transactions.html
  license of original test:
    " Any copyright is dedicated to the Public Domain.
      http://creativecommons.org/publicdomain/zero/1.0/ "
-->
<html>
<head>
<link rel="stylesheet" href="../../../fast/js/resources/js-test-style.css">
<script src="../../../fast/js/resources/js-test-pre.js"></script>
<script src="../resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test IndexedDB's readonly transactions");
if (window.layoutTestController)
    layoutTestController.waitUntilDone();

function test()
{
    indexedDB = evalAndLog("indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;");
    shouldBeFalse("indexedDB == null");
    IDBDatabaseException = evalAndLog("IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;");
    shouldBeFalse("IDBDatabaseException == null");

    name = window.location.pathname;
    description = "My Test Database";
    request = evalAndLog("indexedDB.open(name, description)");
    request.onsuccess = openSuccess;
    request.onerror = unexpectedErrorCallback;
}

function openSuccess()
{
    db = evalAndLog("db = event.target.result");

    request = evalAndLog("request = db.setVersion('1')");
    request.onsuccess = cleanDatabase;
    request.onerror = unexpectedErrorCallback;
}

function cleanDatabase()
{
    deleteAllObjectStores(db);

    osName = "test store";
    objectStore = evalAndLog("objectStore = db.createObjectStore(osName, { autoIncrement: true });");
    evalAndExpectException("db.transaction([osName]).objectStore(osName).add({});", "IDBDatabaseException.READ_ONLY_ERR");
    evalAndExpectException("db.transaction(osName).objectStore(osName).add({});", "IDBDatabaseException.READ_ONLY_ERR");
    key1 = evalAndLog("key1 = 1;");
    key2 = evalAndLog("key2 = 2;");
    evalAndExpectException("db.transaction([osName]).objectStore(osName).put({}, key1);", "IDBDatabaseException.READ_ONLY_ERR");
    evalAndExpectException("db.transaction(osName).objectStore(osName).put({}, key2);", "IDBDatabaseException.READ_ONLY_ERR");
    evalAndExpectException("db.transaction([osName]).objectStore(osName).put({}, key1);", "IDBDatabaseException.READ_ONLY_ERR");
    evalAndExpectException("db.transaction(osName).objectStore(osName).put({}, key1);", "IDBDatabaseException.READ_ONLY_ERR");
    evalAndExpectException("db.transaction([osName]).objectStore(osName).delete(key1);", "IDBDatabaseException.READ_ONLY_ERR");
    evalAndExpectException("db.transaction(osName).objectStore(osName).delete(key2);", "IDBDatabaseException.READ_ONLY_ERR");
    done();
}

var successfullyParsed = true;

test();

</script>
</body>
</html>
